+++ packages/analytics/analytics-chart/src/components/chart-types/StackedBarChart.vue
@@ -145,7 +145,7 @@ const SCROLL_MAX = 10
 const AXIS_BOTTOM_OFFSET = 10
 const AXIS_RIGHT_PADDING = 1
 
-const chartContainerRef = ref<HTMLDivElement| null>(null)
+const chartContainerRef = ref<HTMLDivElement | null>(null)
 
 const totalValueOfDataset = ({ chart }: EventContext, label: string) => {
   const chartData: BarChartData = chart.data as BarChartData