+++ packages/analytics/analytics-metric-provider/src/types/fetcher-types.ts
@@ -1,10 +1,11 @@
 import type {
   Timeframe,
-  ExploreAggregations, QueryableExploreDimensions, ExploreFilter, AnalyticsBridge,
+  ExploreAggregations, QueryableExploreDimensions, ExploreFilter, AnalyticsBridge, QueryDatasource,
 } from '@kong-ui-public/analytics-utilities'
 import type { Ref } from 'vue'
 
 export interface MetricFetcherOptions {
+  datasource: Ref<QueryDatasource>
   metrics: Ref<ExploreAggregations[]>
   dimensions?: QueryableExploreDimensions[]
   filter: Ref<ExploreFilter[] | undefined>