+++ packages/analytics/analytics-config-store/CHANGELOG.md
@@ -3,6 +3,45 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@0.7.0...@kong-ui-public/analytics-config-store@0.7.1) (2024-06-14)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
+# [0.7.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@0.6.1...@kong-ui-public/analytics-config-store@0.7.0) (2024-06-13)
+
+
+### Features
+
+* **analytics:** add support for basic datasource [MA-2931, MA-2923] ([#1450](https://github.com/Kong/public-ui-components/issues/1450)) ([6b5e367](https://github.com/Kong/public-ui-components/commit/6b5e367d0bf7f2f4c12117e30280ffcf819bfd68))
+
+
+### BREAKING CHANGES
+
+* **analytics:** change signature of `queryFn` in query bridge
+
+- Change signature of `queryFn`.
+- Update default queried time range to 7 days for everyone behind a feature flag.
+- Stop relying on trend access determination in metric cards.
+- Add `datasource` parameter to MetricsProvider.
+- Add `datasource` property to dashboard query definitions.
+- Determine a default `datasource` based on feature flag.
+
+
+
+
+
+## [0.6.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@0.6.0...@kong-ui-public/analytics-config-store@0.6.1) (2024-06-10)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
 # [0.6.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@0.5.1...@kong-ui-public/analytics-config-store@0.6.0) (2024-06-06)