+++ packages/core/cli/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/cli@2.1.2...@kong-ui-public/cli@2.1.3) (2024-06-11)
+
+**Note:** Version bump only for package @kong-ui-public/cli
+
+
+
+
+
+## [2.1.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/cli@2.1.1...@kong-ui-public/cli@2.1.2) (2024-06-11)
+
+**Note:** Version bump only for package @kong-ui-public/cli
+
+
+
+
+
+## [2.1.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/cli@2.1.0...@kong-ui-public/cli@2.1.1) (2024-06-07)
+
+**Note:** Version bump only for package @kong-ui-public/cli
+
+
+
+
+
 # [2.1.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/cli@2.0.11...@kong-ui-public/cli@2.1.0) (2024-06-06)